log/SubsystemMap: return a size_t from get_num

+++ src/log/SubsystemMap.h
@@ -28,7 +28,7 @@ class SubsystemMap {
 public:
   SubsystemMap() : m_max_name_len(0) {}
 
-  int get_num() const {
+  size_t get_num() const {
     return m_subsys.size();
   }
